os_uptime() takes no params, returns uptime in usec.

+++ src/os/time.h
@@ -7,8 +7,8 @@ struct timer {
     struct sched_task *task;
 };
 
-extern double get_current_time();
-extern double get_init_time();
+extern long get_current_time();
+extern long get_init_time();
 
 int set_timer(int seconds);
 extern struct timer *new_timer(int seconds, struct sched_task *task, struct timer *tmr);
